+++ src/masks.py
@@ -1,7 +1,7 @@
 """Функция реализует маскировки счета"""
 
 
-def get_mask_account(personal_account: int) ->str:
+def get_mask_account(personal_account: int) -> str:
     personal_account_str = str(personal_account)
     if len(personal_account_str) == 20:
         nomber_card_number = f"**{personal_account_str[-4:]}"
@@ -14,7 +14,7 @@ def get_mask_account(personal_account: int) ->str:
 """Функция реализует маскировки номера"""
 
 
-def get_mask_card_number(personal_card: int) ->str:
+def get_mask_card_number(personal_card: int) -> str:
     personal_card_str = str(personal_card)
     if len(personal_card_str) == 16:
         nomber_card_numbe = f"{personal_card_str[:4]} {personal_card_str[4:6]} ** *** {personal_card_str[12:]}"
