--- test/test_processing.py
@@ -1,34 +1,0 @@
-import pytest
-from src.processing import filter_by_state,sort_by_date
-
-@pytest.mark.parametrize("h , h1" , [
-    ([{"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},{"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},{"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},{"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"}] , [{'id': 41428829, 'state': 'EXECUTED', 'date': '2019-07-03T18:35:29.512364'}, {'id': 939719570, 'state': 'EXECUTED', 'date': '2018-06-30T02:08:58.425572'}]),
-    ([{"id": 41428829, "state": "CANCELED", "date": "2019-07-03T18:35:29.512364"},{"id": 939719570, "state": "CANCELED", "date": "2018-06-30T02:08:58.425572"},{"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},{"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"}] , [])
-])
-
-
-def test_filter_by_state(h , h1):
-    assert filter_by_state(h) == h1
-
-def test_filter_by():
-    assert filter_by_state([{"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-                            {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-                            {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
-                            {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"}],
-                           "EXECUTED") == ([{"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-                                            {"id": 939719570, "state": "EXECUTED",
-                                             "date": "2018-06-30T02:08:58.425572"}])
-
-@pytest.mark.parametrize("g , g1", [
-    ([{"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-     {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-     {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
-     {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"}],
-     [{'id': 41428829, 'state': 'EXECUTED', 'date': '2019-07-03T18:35:29.512364'},
-      {'id': 615064591, 'state': 'CANCELED', 'date': '2018-10-14T08:21:33.419441'},
-      {'id': 594226727, 'state': 'CANCELED', 'date': '2018-09-12T21:27:25.241689'},
-      {'id': 939719570, 'state': 'EXECUTED', 'date': '2018-06-30T02:08:58.425572'}]),
-
-])
-def test_filter_1(g , g1):
-    assert sort_by_date(g) == g1
