+++ tests_1/test_masks.py
@@ -1,5 +1,6 @@
 import pytest
-from src.widget import get_mask_card_number, get_mask_account
+
+from src.masks import get_mask_account, get_mask_card_number
 
 
 @pytest.mark.parametrize(
