+++ tests_1/test_widget.py
@@ -1,5 +1,6 @@
-from src.widget import get_date, mask_account_card
 import pytest
+
+from src.widget import get_date, mask_account_card
 
 
 def test_get_date():
